Add doctor management and generic time utilities
This commit introduces administrative functions for managing doctors, such as creating new doctor entries and deleting existing ones. These features were added to the `doctors.vue` admin page. Additionally, a generic date and time formatting function was created in `formatDateTime.js`. The dashboard for the admin was also updated and a new function `useDashboard` was added to aid in the display of information. The commit also included the creation and modification of the layouts for these new features. Finally, necessary updates were made to yarn dependencies.