Update signup form validation and enhance UX
This commit introduces TypeScript in the the SignUp form and implements form schema validation using 'zod' library. Fields in the SignUp form now have detailed validations to guide the user when inputting information. On submission, 'FormSubmitEvent' captures the values from the form to process the request. An error message display was also added to inform users on any encountered error. The features improve the form's user experience and prevent submission of incorrect data. A logo image was also added to the main navbar and responsive layouts adopted for improved user interface across different devices.