Refactor default layout and add sign-out functionality
The default.vue layout was significantly refactored to include new tags and elements. This includes the replacement of the Navbar with a more complex UHeader and the addition of a sign-out functionality. The sign-out function makes use of the authStore to reset user data and then redirects to the login page. Furthermore, commented-out code and debugging logs were added, which will likely be used for future development and debugging.