Refactor and streamline authentication form logic
This commit significantly refactors the signup and login forms in order to simplify their code structure and enhance maintainability. The forms are reduced by replacing manual fields and logic with a standardized UAuthForm component that handles all the form fields dynamically. The 'isMobile' checks are also removed from the pages and forms, to make the design more responsive. Error handling is also improved in the updated logic.