Replace landing card with NuxtLink for location directions
This update switches the ULandingCard component containing the location directions with a NuxtLink, effectively creating a direct, external link to the map location of the establishment. This minor but crucial modification improves user navigation experience by providing a quick and easy way to access the location information externally on google maps from the website, instead of displaying maps within the page. The code has also been cleaned up to remove an unnecessary image.