Update Instagram button URL and refactor homepage elements
Code changes include the update of the destination URL for an Instagram button on the default layout page, and multiple updates to the homepage structure for better readability and user experience. This includes the refactoring of the ULandingCard component and rearrangement of elements like the LandingFaqAccordion and the location link, which now is a NuxtLink for a better redirection user experience.