Update admin pages and implement search filter
This commit reorganizes the layout on the admin dashboard and introduces a search filter function on users, doctors, and appointments pages. Various visual elements on the admin pages have been updated for better UI. The code also includes the addition of an 'admin' middleware to handle route transitions based on user authentication and role. Finally, a designated function was added to handle elevation of user privileges.